refactor: Move RebindSafeMBeanServer to presto-common
Move RebindSafeMBeanServer from individual connector modules to
presto-common/util to eliminate code duplication across 5 connectors.
