+++ presto-cassandra/src/main/java/com/facebook/presto/cassandra/CassandraConnectorFactory.java
@@ -15,6 +15,7 @@ package com.facebook.presto.cassandra;
 
 import com.facebook.airlift.bootstrap.Bootstrap;
 import com.facebook.airlift.json.JsonModule;
+import com.facebook.presto.common.util.RebindSafeMBeanServer;
 import com.facebook.presto.spi.ConnectorHandleResolver;
 import com.facebook.presto.spi.connector.Connector;
 import com.facebook.presto.spi.connector.ConnectorContext;
