+++ presto-jmx/src/main/java/com/facebook/presto/connector/jmx/JmxConnectorFactory.java
@@ -14,7 +14,7 @@
 package com.facebook.presto.connector.jmx;
 
 import com.facebook.airlift.bootstrap.Bootstrap;
-import com.facebook.presto.connector.jmx.util.RebindSafeMBeanServer;
+import com.facebook.presto.common.util.RebindSafeMBeanServer;
 import com.facebook.presto.spi.ConnectorHandleResolver;
 import com.facebook.presto.spi.NodeManager;
 import com.facebook.presto.spi.connector.Connector;
