+++ presto-thrift-connector/src/main/java/com/facebook/presto/connector/thrift/ThriftConnectorFactory.java
@@ -16,7 +16,7 @@ package com.facebook.presto.connector.thrift;
 import com.facebook.airlift.bootstrap.Bootstrap;
 import com.facebook.drift.transport.netty.client.DriftNettyClientModule;
 import com.facebook.presto.common.type.TypeManager;
-import com.facebook.presto.connector.thrift.util.RebindSafeMBeanServer;
+import com.facebook.presto.common.util.RebindSafeMBeanServer;
 import com.facebook.presto.spi.ConnectorHandleResolver;
 import com.facebook.presto.spi.ConnectorSystemConfig;
 import com.facebook.presto.spi.connector.Connector;
